+++ src/algebra/layers.jl
@@ -1,3 +1,9 @@
+"""
+    Layers(layers::Vector{Layer})
+
+Algebraic object encoding a list of [`AlgebraOfGraphics.Layer`](@ref) objects.
+`Layers` objects can be added or multiplied, yielding a novel `Layers` object.
+"""
 struct Layers
     layers::Vector{Layer}
 end
@@ -132,7 +138,7 @@ function compute_axes_grid(fig, s::OneOrMoreLayers;
             label = compute_label(ae.entries, i)
             for (k′, v) in pairs((label=string(label), ticks=ticks(scale)))
                 k = Symbol(var, k′)
-                k in keys(axis) || (getproperty(Axis(ae), k)[] = v)
+                k in keys(axis) || (getproperty(ae.axis, k)[] = v)
             end
         end
     end
@@ -153,16 +159,33 @@ function Makie.plot(s::OneOrMoreLayers;
     return FigureGrid(fig, grid)
 end
 
-# Convenience function, which may become superfluous if `plot` also calls `facet!`
+"""
+    draw(s; axis=NamedTuple(), figure=NamedTuple, palettes=NamedTuple())
+
+Draw a [`AlgebraOfGraphics.Layer`](@ref) or [`AlgebraOfGraphics.Layers`](@ref) object `s`.
+The output can be customized by giving axis attributes to `axis`, figure attributes
+to `figure`, or custom palettes to `palettes`.
+Legend and colorbar are drawn automatically. For finer control, use [`draw!`](@ref),
+[`legend!`](@ref), and [`colorbar!`](@ref) independently.
+"""
 function draw(s::OneOrMoreLayers;
               axis = NamedTuple(), figure=NamedTuple(), palettes=NamedTuple())
     fg = plot(s; axis, figure, palettes)
     facet!(fg)
-    Legend(fg)
+    colorbar!(fg)
+    legend!(fg)
     resizetocontent!(fg)
     return fg
 end
 
+"""
+    draw!(fig, s; axis=NamedTuple(), palettes=NamedTuple())
+
+Draw a [`AlgebraOfGraphics.Layer`](@ref) or [`AlgebraOfGraphics.Layers`](@ref) object `s` on `fig`.
+`fig` can be a figure, a position in a layout, or an axis if `s` has no facet specification.
+The output can be customized by giving axis attributes to `axis` or custom palettes
+to `palettes`.  
+"""
 function draw!(fig, s::OneOrMoreLayers;
                axis=NamedTuple(), palettes=NamedTuple())
     ag = plot!(fig, s; axis, palettes)